iglooChatBackend/requirements.txt
djangorestframework
django-allauth
dj-rest-auth
django-cors-headers
requests